Move pkg to std/
