+++ codec/std/codec.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-package cosmos_sdk.simapp.codec.v1;
+package cosmos_sdk.codec.std.v1;
 
 import "third_party/proto/cosmos-proto/cosmos.proto";
 import "third_party/proto/gogoproto/gogo.proto";
@@ -12,7 +12,7 @@ import "x/params/types/proposal/types.proto";
 import "x/upgrade/types/types.proto";
 import "x/distribution/types/types.proto";
 
-option go_package = "github.com/cosmos/cosmos-sdk/simapp/codec";
+option go_package = "github.com/cosmos/cosmos-sdk/codec/std";
 
 // Account defines the application-level Account type.
 message Account {
